Fix user being deleted when it shouldn't (fix #273)

+++ openrepairplatform/event/views/event_participants.py
@@ -156,7 +156,8 @@ class CancelReservationView(RedirectView):
         event.registered.remove(user)
         event.presents.remove(user)
 
-        if user.first_name == "":
+        # Check if user is registered to others events
+        if user.first_name == "" and not (user.registered_events.count() > 0):
             # This is a temporary user created only for this event, we can
             # delete it
             user.delete()
